refactor: Adjust package name
