+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/player/hide/HidePlayerButtonsPatch.kt
@@ -2,7 +2,7 @@ package app.revanced.patches.youtube.layout.buttons.player.hide
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patches.youtube.layout.buttons.player.HidePlayerOverlayButtonsPatch
+import app.revanced.patches.youtube.layout.buttons.overlay.HidePlayerOverlayButtonsPatch
 
 @Deprecated("This patch has been merged into HidePlayerOverlayButtonsPatch.")
 object HidePlayerButtonsPatch : BytecodePatch(
